Accept comma-separated list for function declarations.
The declaration of a function without prototype is currently
silently accepted by sparse but a warning is issued for
'old-style' declarations:
        ... warning: non-ANSI function declaration ...

However, the difference between these two cases is made by
checking if a ';' directly follow the parentheses. So:
        int foo();
is silently accepted, while a warning is issued for:
        int foo(a) int a;
but also for:
        int foo(), bar();
This last case, while unusual, is not less ANSI than a simple
'int foo();'. It's just detected so because there is no ';'
directly after the first '()'.

Fix this by also using ',' to detect the end of function
declarations and their ANSIness.

(From upstream sparse: a0f34e0fd2a44016ef668e19187a342c5f4edf45)
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>