constant: add -Wconstant-suffix warning
Currently, when used on the kernel, sparse issues a bunch
of warnings like:
        warning: constant 0x100000000 is so big it is long

These warning are issued when there is a discrepancy
between the type as indicated by the suffix (or the absence
of a suffix) and the real type as selected by the type
suffix *and* the value of the constant.

Since there is nothing incorrect with this discrepancy,
(no bits are lost) these warnings are more annoying than useful.
So, make them depending on a new warning flag -Wconstant-suffix
and make it off by default.

(From sparse upstream fdf8252f312a40df9aa51c6e30c0d07fa29ebd12)
Signed-off-by: Ramsay Jones <ramsay@ramsayjones.plus.com>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>